requirements.txt
Django>=5.2
PyPDF2>=3.0.0
scikit-learn>=1.0.0
nltk>=3.8.1
google-generativeai>=0.3.0

# Note: After installing, run 'python download_nltk_resources.py' to download required NLTK data 